requirements.txt
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.3
markdown
pytz==2017.2
